+++ app/models/disqualified/sequence_record.rb
@@ -1,0 +1,5 @@
+# typed: strict
+
+class Disqualified::SequenceRecord < Disqualified::BaseRecord
+  self.table_name = "disqualified_sequences"
+end