+++ lib/generators/disqualified/install/install_generator.rb
@@ -3,6 +3,9 @@ class Disqualified::InstallGenerator < Rails::Generators::Base
 
   def copy_migration_file
     basename = "20220703062536_create_disqualified_jobs.rb"
-    copy_file basename, "db/migrate/#{basename}"
+    copy_file(basename, "db/migrate/#{basename}")
+
+    basename = "20241119023328_create_disqualified_sequences.rb"
+    copy_file(basename, "db/migrate/#{basename}")
   end
 end