+++ CHANGELOG.md
@@ -7,6 +7,10 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## Unreleased
 
+### Added
+
+* Added `Disqualified::Sequence`
+
 ### Changed
 
 * Renamed `Disqualified::Record#unqueue` to `Disqualified::Record#unclaim`